Create README.md
For a classiﬁcation of handwritten digits : A well known benchmark database that is extensively used in the literature is the MNIST, which contains a large number of handwritten digits, centered and normalized into ﬁxed-size images (28×28 pixels). Here we build a K Nearest Neighbor Classifier to recognize digits. 

Language : python

IDE: ipython notebook

Author: Tanay Shankar

2018